Game settings (#16)
* reorder ui services

* add game settings modal

* impl basic settings modal

* impl basic settings page

* add config for maps usage

* add maps control

* add missing usings

* get maps working

* basic permissions for settings

* add better init logic for page

* add api to compose

* move file in subfolder

* basic impl for game area selection

* update room settings on geofence changed

* push local changes for switch to v2

+++ TagGame.Shared/Domain/Common/Location.cs
@@ -1,4 +1,4 @@
 ﻿namespace TagGame.Shared.Domain.Common;
 
-public record struct Location(decimal Latitude, decimal Longitude);
+public record struct Location(double Latitude, double Longitude);
